feat: Refactor monitoring inspect response
